Move states to the separate folders
