+++ client/src/app/about/about.module.js
@@ -1,4 +1,4 @@
-import states from './states/states.config';
+import states from './states';
 import uiRouter from 'angular-ui-router';
 
 export default angular.module('app.about', [uiRouter])
