+++ client/src/app/contacts/contacts.module.js
@@ -3,7 +3,7 @@ import angularMessages from 'angular-messages';
 import angularResource from 'angular-resource';
 import appCommonsModule from '../commons/commons.module';
 import { contactFactory } from './services';
-import states from './states/states.config';
+import states from './states';
 import uiRouter from 'angular-ui-router';
 
 export default angular.module('app.contacts', [
